updating unit tests
